add support for passing custom ethclient to Seth (including Simulated Backend)
