+++ seth/config.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/ethclient/simulated"
 	"github.com/pelletier/go-toml/v2"
 	"github.com/pkg/errors"
 )
@@ -45,6 +46,7 @@ type Config struct {
 	revertedTransactionsFile string
 	ephemeral                bool
 	RPCHeaders               http.Header
+	ethclient                simulated.Client
 
 	// external fields
 	// ArtifactDir is the directory where all artifacts generated by seth are stored (e.g. transaction traces)
@@ -263,11 +265,25 @@ func (c *Config) Validate() error {
 		c.PendingNonceProtectionTimeout = &Duration{D: DefaultPendingNonceProtectionTimeout}
 	}
 
+	if c.ethclient == nil && len(c.Network.URLs) == 0 {
+		return errors.New("at least one url should be present in config in 'secret_urls = []'")
+	}
+
+	if c.ethclient != nil && len(c.Network.URLs) > 0 {
+		return errors.New(EthClientAndUrlsSet)
+	}
+
 	return nil
 }
 
-// FirstNetworkURL returns first network URL
-func (c *Config) FirstNetworkURL() string {
+// MustFirstNetworkURL returns first network URL or panics if it's not set
+func (c *Config) MustFirstNetworkURL() string {
+	if c.Network == nil {
+		panic("no Network is set in the config")
+	}
+	if len(c.Network.URLs) == 0 {
+		panic("no URLs are set in the Network config")
+	}
 	return c.Network.URLs[0]
 }
